Merged PR 1483: Extract methods and refactor current code to increase readability in StepService
Refactor code in StepService class and add two tests to StepServiceTest. Changes in other classes are only in form of method name- changes.

Add two lists two make the use of Optional-wrappers easier. Not the most clean solution but will do for now? Will probably be easy to replace when a Mapper is introduced.

Related work items: #2648, #2649